+++ parser_library/src/preprocessor_options.h
@@ -49,7 +49,7 @@ struct cics_preprocessor_options
     bool operator==(const cics_preprocessor_options&) const = default;
 };
 
-using preprocessor_options = std::variant<std::monostate, db2_preprocessor_options, cics_preprocessor_options>;
+using preprocessor_options = std::variant<db2_preprocessor_options, cics_preprocessor_options>;
 
 } // namespace hlasm_plugin::parser_library
 #endif