+++ parser_library/src/workspaces/processor_file_impl.cpp
@@ -45,8 +45,10 @@ void processor_file_impl::collect_diags() const { file_impl::collect_diags(); }
 
 bool processor_file_impl::is_once_only() const { return false; }
 
-parse_result processor_file_impl::parse(
-    parse_lib_provider& lib_provider, asm_option asm_opts, preprocessor_options pp, virtual_file_monitor* vfm)
+parse_result processor_file_impl::parse(parse_lib_provider& lib_provider,
+    asm_option asm_opts,
+    std::vector<preprocessor_options> pp,
+    virtual_file_monitor* vfm)
 {
     if (!last_analyzer_opencode_)
         last_opencode_id_storage_ = std::make_shared<context::id_storage>();