+++ parser_library/test/workspace/processor_group_test.cpp
@@ -27,9 +27,15 @@ auto asm_options(config::assembler_options o)
     processor_group("", std::move(o), {}).update_asm_options(result);
     return result;
 }
-auto pp_options(decltype(config::preprocessor_options::options) o)
+
+template<typename... Opts>
+auto pp_options(Opts... o)
 {
-    return processor_group("", {}, config::preprocessor_options { .options = std::move(o) }).preprocessor();
+    return processor_group("",
+        {},
+        { config::preprocessor_options {
+            .options = decltype(config::preprocessor_options::options)(std::move(o)) }... })
+        .preprocessors();
 }
 
 TEST(processor_group, assembler_options)
@@ -51,16 +57,20 @@ TEST(processor_group, preprocessor_options)
     constexpr auto cics = [](auto... p) { return preprocessor_options(cics_preprocessor_options(std::move(p)...)); };
     constexpr auto db2 = [](auto... p) { return preprocessor_options(db2_preprocessor_options(std::move(p)...)); };
 
-    EXPECT_EQ(pp_options(std::monostate()), preprocessor_options());
+    EXPECT_EQ(pp_options(), std::vector<preprocessor_options> {});
 
-    EXPECT_EQ(pp_options(config::db2_preprocessor {}), db2());
-    EXPECT_EQ(pp_options(config::db2_preprocessor { .version = "A" }), db2("A"));
+    EXPECT_EQ(pp_options(config::db2_preprocessor {}), std::vector { db2() });
+    EXPECT_EQ(pp_options(config::db2_preprocessor { .version = "A" }), std::vector { db2("A") });
 
-    EXPECT_EQ(pp_options(config::cics_preprocessor {}), cics());
-    EXPECT_EQ(pp_options(config::cics_preprocessor { .leasm = true }), cics(true, true, true));
-    EXPECT_EQ(pp_options(config::cics_preprocessor { .prolog = false }), cics(false, true, false));
-    EXPECT_EQ(pp_options(config::cics_preprocessor { .epilog = false }), cics(true, false, false));
-    EXPECT_EQ(pp_options(config::cics_preprocessor { .prolog = false, .leasm = true }), cics(false, true, true));
+    EXPECT_EQ(pp_options(config::cics_preprocessor {}), std::vector { cics() });
+    EXPECT_EQ(pp_options(config::cics_preprocessor { .leasm = true }), std::vector { cics(true, true, true) });
+    EXPECT_EQ(pp_options(config::cics_preprocessor { .prolog = false }), std::vector { cics(false, true, false) });
+    EXPECT_EQ(pp_options(config::cics_preprocessor { .epilog = false }), std::vector { cics(true, false, false) });
+    EXPECT_EQ(pp_options(config::cics_preprocessor { .prolog = false, .leasm = true }),
+        std::vector { cics(false, true, true) });
+
+    EXPECT_EQ(pp_options(config::cics_preprocessor { .leasm = true }, config::db2_preprocessor { .version = "A" }),
+        (std::vector<preprocessor_options> { cics(true, true, true), db2("A") }));
 }
 
 TEST(processor_group, asm_options_optable_valid)